Changing IntoWasmAbi to use interning
