+++ src/lib.rs
@@ -57,6 +57,8 @@ pub mod prelude {
     }
 }
 
+#[allow(unused)]
+mod cache;
 pub mod convert;
 pub mod describe;
